backend/requirements.txt
fastapi
uvicorn
psycopg2-binary
sqlalchemy
alembic
passlib[bcrypt]
openai>=1.0.0
python-dotenv
python-multipart
pydantic
jinja2
python-jose[cryptography]
pydantic[email]
python-magic
python-docx
PyPDF2

# 🔧 Дополнительные зависимости для продакшена
redis>=4.0.0
psutil>=5.8.0
boto3>=1.26.0
requests>=2.28.0
aiofiles>=0.8.0
httpx>=0.24.0

# 📊 Мониторинг и метрики
prometheus-client>=0.16.0

# 🗄️ Работа с БД и бэкапами
# pg-dump>=0.1.0  # Эта зависимость недоступна, используйте системный pg_dump 